% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TallaridaFct.R
\name{AdditiveIsobole}
\alias{AdditiveIsobole}
\title{AdditiveIsobole() Function}
\usage{
AdditiveIsobole(effect, sameB = "T", dataB, IC50A, IC50B, mA, mB, rho)
}
\arguments{
\item{effect}{: Effects that additive isobole produce (Often 50)}

\item{sameB}{: higher potency dose is equal to dose B}

\item{dataB}{: Dose Effect data of higher potency dose}

\item{IC50A}{: Relative IC50 of dose A(less potency dose)}

\item{IC50B}{: Relative IC50 of dose B(higher potency dose)}

\item{mA}{: slope of dose A(less potency dose)}

\item{mB}{: slope of dose B(higher potency dose)}

\item{rho}{: rho obtained from CalculateRho()}
}
\description{
This function calculates the additive isobole from Tallarida null model.
}
\examples{
AdditiveIsobole()
}
\keyword{Additive}
\keyword{Tallaridamodel,}
\keyword{isobole}
