% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationFct.R
\name{CalculateDose}
\alias{CalculateDose}
\title{CalculateDose() Function}
\usage{
CalculateDose(data, Lambda)
}
\arguments{
\item{data}{: estimated dose of the effect}

\item{Lambda}{: fixed mixture ratio (doseA)}
}
\description{
This function divide the estimated dose of combined data into dose A and dose B by taking lambda.
}
\examples{
CalculateDose()
}
\keyword{Estimate}
\keyword{combined}
\keyword{doses}
