% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TallaridaFct.R
\name{CalculRho}
\alias{CalculRho}
\title{CalculRho() Function}
\usage{
CalculRho(EmindoseA, EmaxdoseA, EmindoseB, EmaxdoseB)
}
\arguments{
\item{EmindoseA}{: lower aymptote of less potency dose}

\item{EmaxdoseA}{: upper asymptote of less potency dose}

\item{EmindoseB}{: lower aymptote of higher potency dose}

\item{EmaxdoseB}{:upper asymptote of higher potency dose}
}
\description{
This function calculates the necessary rho for the equation of the additive isobole.
}
\examples{
CalculRho()
}
\keyword{Additive}
\keyword{CalculateRho,}
\keyword{Tallaridamodel,}
\keyword{isobole}
