% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationFct.R
\name{CombinationData}
\alias{CombinationData}
\title{CombinationData() Function}
\usage{
CombinationData(paramC, Rep, sd, lambda, ray)
}
\arguments{
\item{paramC}{: Vectors containing parameters(Emin, Emax, m, IC50)}

\item{sd}{: sd}

\item{lambda}{: fixed mixture ratio of dose A (Ray)}

\item{ray}{: following ray}

\item{replicate}{: number of replicates}
}
\value{
I, Replicate, Comb, doseA, doseB, Ray, lambda
}
\description{
This function does simulation with combination data.
}
\examples{
CombinationData()

}
\keyword{Simulation}
