% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFct.R
\name{CountAdditivity}
\alias{CountAdditivity}
\title{CountAdditivity() Function}
\usage{
CountAdditivity(Estfinal, Additivity, nameModel)
}
\arguments{
\item{Estfinal}{: estimated data}

\item{Additivity}{: Additivity data obtained with null model}

\item{nameModel}{: name of the null model}
}
\description{
This function count the combination synergy of the estimated point
}
\examples{
CountAdditivity()

}
\keyword{Additivity}
\keyword{count}
