% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFct.R
\name{CountAdditivity_Tallarida}
\alias{CountAdditivity_Tallarida}
\title{CountAdditivity_Tallarida() Function}
\usage{
CountAdditivity_Tallarida(Estfinal, BtoA, AtoB, nameModel)
}
\arguments{
\item{Estfinal}{: estimated data}

\item{BtoA}{: Additivity Tallarida B to A}

\item{AtoB}{: Additivity Tallarida A to B}

\item{nameModel}{: name of the null model}
}
\description{
This function count the combination synergy of the estimated point for Tallarida model (twolines)
}
\examples{
CountAdditivity_Tallarida()

}
\keyword{Additivity}
\keyword{count}
