% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoseeffectFct.R
\name{DoseEffect}
\alias{DoseEffect}
\title{DoseEffect() Function}
\usage{
DoseEffect(x, Emin, Emax, m, IC50)
}
\arguments{
\item{x}{: amount of dose}

\item{Emin}{: lower asymptote}

\item{Emax}{: upper asymptote}

\item{m}{: slope (Hill's coefficient)}

\item{IC50}{: Relative IC50}
}
\value{
Efffect that is produced by x amount of doses
}
\description{
This function produces effects from dose-effect 4P logistic parameter model.
}
\examples{
DoseEffect()
}
\keyword{4P}
\keyword{DoseEffect,}
\keyword{logistic}
\keyword{model}
