% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationFct.R
\name{EstimateParameters}
\alias{EstimateParameters}
\title{EstimateParameters() Function}
\usage{
EstimateParameters(data, nameA, nameB)
}
\arguments{
\item{data}{: amount of dose}

\item{nameA}{: name of dose A}

\item{nameB}{: name of dose B}
}
\description{
This function estimates the parameters of doseA and doseB from data.
}
\examples{
EstimateParameters()
}
\keyword{(free}
\keyword{4P}
\keyword{Estimate}
\keyword{data,}
\keyword{from}
\keyword{logistic}
\keyword{model}
\keyword{model)}
\keyword{parameters}
