% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TallaridaFct.R
\name{EstimateParametersTallarida}
\alias{EstimateParametersTallarida}
\title{EstimateParametersTallarida() Function}
\usage{
EstimateParametersTallarida(data, nameA, nameB, Nexperiment, Nray)
}
\description{
EstimateParametersTallarida() Function
}
\examples{
EstimateParametersTallarida()
}
\keyword{Estimate}
\keyword{combined}
\keyword{doses}
