% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationFct.R
\name{EstimatePE}
\alias{EstimatePE}
\title{EstimatePE() Function}
\usage{
EstimatePE(data, nameA, nameB, Nexperiment, Nray)
}
\arguments{
\item{data}{: estimated dose of the effect}
}
\description{
EstimatePE() Function
}
\examples{
EstimatePE()

}
\keyword{Estimate}
\keyword{combined}
\keyword{doses}
