% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationFct.R
\name{EstimateValue}
\alias{EstimateValue}
\title{EstimateValue() Function}
\usage{
EstimateValue(data, nameA, nameB, Nexperiment, Nray, ICvalue)
}
\arguments{
\item{data}{: amount of dose}

\item{nameA}{: name of dose A}

\item{nameB}{: name of dose B}

\item{Nexperiment}{: total number of experiments}

\item{Nray}{: total number of rays}

\item{ICvalue}{: effect}
}
\description{
This function estimates combined doses of each effect from data.
}
\examples{
EstimateValue()
}
\keyword{4P}
\keyword{Estimate}
\keyword{combined}
\keyword{doses,}
\keyword{logistic}
\keyword{model}
