% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFct.R
\name{EvaluateAdditivity}
\alias{EvaluateAdditivity}
\title{EvaluateAdditivity() Function}
\usage{
EvaluateAdditivity(Estfinal, Additivity, nameModel)
}
\arguments{
\item{Estfinal}{: Estimated data}

\item{Additivity}{: Additivity data obtained with null model}

\item{nameModel}{: name of the null model}
}
\description{
This function evaluate the combination synergy of the estimated point
}
\examples{
EvaluateAdditivity()

}
\keyword{Additivity}
\keyword{Evaluate}
