% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFct.R
\name{EvaluateAdditivity_Tallarida}
\alias{EvaluateAdditivity_Tallarida}
\title{EvaluateAdditivity_Tallarida() Function}
\usage{
EvaluateAdditivity_Tallarida(Estfinal, BtoA, AtoB, nameModel)
}
\arguments{
\item{Estfinal}{: estimated data}

\item{BtoA}{: Additivity data obtained with Tallarida B to A}

\item{AtoB}{: Additivity data obtained with Tallarida A to B}

\item{nameModel}{: name of the null model}
}
\description{
This function evaluate the combination synergy of the estimated point using Tallarida model (two lines)
}
\examples{
EvaluateAdditivity_Tallarida()

}
\keyword{Additivity}
\keyword{Evaluate}
