% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationFct.R
\name{ExtractParameters}
\alias{ExtractParameters}
\title{ExtractParameters() Function}
\usage{
ExtractParameters(paramA, paramB)
}
\arguments{
\item{paramA}{: parameters of compound A (Emin, Emax, m, IC50rel(exponential value))}

\item{paramb}{:parameters of compound B (Emin, Emax, m, IC50rel(exponential value))}
}
\description{
ExtractParameters() Function
}
\examples{
ExtractParameters()

}
\keyword{Combine}
\keyword{compoundsinto}
\keyword{dataset}
\keyword{of}
\keyword{one}
\keyword{parameters}
\keyword{single}
