% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFct.R
\name{FinalAdditivity}
\alias{FinalAdditivity}
\title{FinalAdditivity() Function}
\usage{
FinalAdditivity(Estfinal, Loewe, TallaridaBtoA, TallaridaAtoB, Hand)
}
\arguments{
\item{Estfinal}{: estimated data}

\item{Loewe}{: Additivity Loewe}

\item{TallaridaBtoA}{: Additivity Tallarida B to A}

\item{TallaridaAtoB}{: Additivity Tallarida A to B}

\item{Hand}{: Additivity Hand}

\item{HSA}{: Additivity HSA}
}
\description{
Combine the result obtained from all null models
}
\examples{
FinalAdditivity()

}
\keyword{Additivity}
\keyword{count}
