% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandmodelFct.R
\name{HandData}
\alias{HandData}
\title{HandData() Function}
\usage{
HandData(hand, p, ICvalue)
}
\arguments{
\item{hand}{: parameters obtained by Hand model}

\item{ICvalue}{: effect (often 50)}

\item{paramA}{: parameters of dose B}
}
\description{
This function produces the data from parameters obtained from Hand model
}
\examples{
HandData()
}
\keyword{Handmodel,}
\keyword{Synergy}
