% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandmodelFct.R
\name{Handmodel}
\alias{Handmodel}
\title{Handmodelwithplot() Function}
\usage{
Handmodel(param, lamb, doseeffect = "F", effectsens = "F", returndata = "F")
}
\arguments{
\item{lamb}{: vector of different lambda (not include 0 and 1)}

\item{doseeffect}{: graphical representation of dose-effect curves}

\item{effectsens}{: graphical representation of effect-sensitivity curves}

\item{returndata}{: Return the estimated dose-effect data of the combination. If "F", it returns the parameters.}

\item{p}{: theoretical parameters of doseA and doseB}
}
\description{
This function calculates sensitivity of the dose-effect curve. It is the derivated form of effect from dose-effect function.
}
\examples{
Sensitivity()
}
\keyword{Handmodel,}
\keyword{Synergy}
\keyword{]Sensitivity,}
