% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoseeffectFct.R
\name{ICrelative}
\alias{ICrelative}
\title{ICrelative() Function}
\usage{
ICrelative(Emin, Emax, m, IC50)
}
\arguments{
\item{Emin}{: lower asymptote}

\item{Emax}{: higher axymptote}

\item{m}{: slope (Hill's coefficient)}

\item{IC50}{: Absolute IC50}
}
\value{
Dose is needed to produce x effect
}
\description{
Obtain IC relative
}
\examples{
ICrelative()
}
\keyword{IC50}
\keyword{Relative}
