% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoseeffectFct.R
\name{InverseDoseEffect}
\alias{InverseDoseEffect}
\title{InverseDoseEffect() Function}
\usage{
InverseDoseEffect(x, Emin, Emax, m, IC50)
}
\arguments{
\item{x}{: effect}

\item{Emin}{: lower asymptote}

\item{Emax}{: upper asymptote}

\item{m}{: slope (Hill's coefficient)}

\item{IC50}{: Relative IC50}
}
\value{
Dose is needed to produce x effect
}
\description{
Inversed dose effect function
}
\examples{
InverseDoseEffect()
}
\keyword{4P}
\keyword{InverseDoseEffect,}
\keyword{logistic}
\keyword{model}
