% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoeweFct.R
\name{Loewemodel}
\alias{Loewemodel}
\title{Loewemodel() Function}
\usage{
Loewemodel(ICvalue, param, lambda, Ki = 1)
}
\arguments{
\item{ICvalue}{: Effects (often 50)}

\item{param}{: Theoretical parameters of dose A and doseB}

\item{lambda}{: fixed mixture ratio to print the combinations.}

\item{Ki}{: Combination Index = 1 for additivity.}
}
\description{
This function calculates additivie isoboles by using classical Loewe model. It prints the combination doses with different mixture ratio.
The result only contains doseA, doseB, I. The calculation of doseA and doseB is not influenced by Ki or lambda.
}
\examples{
Loewemodel()

}
\keyword{Additive}
\keyword{Tallaridamodel,}
\keyword{isobole}
