% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoseeffectFct.R
\name{ParameterData}
\alias{ParameterData}
\title{ParameterData() Function}
\usage{
ParameterData(Emin, Emax, m, IC50)
}
\arguments{
\item{Emin}{: lower asymptote}

\item{Emax}{: higher asymptote}

\item{m}{: slope (Hill's coefficient)}

\item{IC50}{: Absolute IC50}
}
\value{
Emin, Emax, m, IC50abs, IC50rel
}
\description{
Returns the parameter data frame
}
\examples{
ParameterData()
}
\keyword{IC50}
\keyword{Relative}
