% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TallaridaFct.R
\name{plotTallarida}
\alias{plotTallarida}
\title{plotTallarida() Function}
\usage{
plotTallarida(isobole)
}
\description{
plotTallarida() Function
}
\examples{
plotTallarida()
}
\keyword{Plot}
\keyword{model}
\keyword{tallarida}
