% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandmodelFct.R
\name{Sensitivity}
\alias{Sensitivity}
\title{Sensitivity() Function}
\usage{
Sensitivity(x, Emin, Emax, m, IC50)
}
\arguments{
\item{x}{: Effect}

\item{Emin}{: lower asymptote}

\item{Emax}{: upper asymptote}

\item{m}{: slope (Hill's coefficient)}

\item{IC50}{: relative IC50}
}
\description{
This function calculates sensitivity of the dose-effect curve. It is the derivated form of effect from dose-effect function.
}
\examples{
Sensitivity()
}
\keyword{Derivate}
\keyword{Effect-sensitivity}
\keyword{curves,}
\keyword{dose}
\keyword{effect}
\keyword{sensitivity,}
