% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationFct.R
\name{SimulateData}
\alias{SimulateData}
\title{SimulateData() Function}
\usage{
SimulateData(param, replicate, sd, doseName)
}
\arguments{
\item{param}{: Vectors containing parameters(Emin, Emax, m, IC50)}

\item{replicate}{: number of replicates}

\item{sd}{: sd}

\item{doseName}{: name of the component}
}
\value{
dose, I, replicate, doseName
}
\description{
This function simulate data with dose = c(0.3, 1, 3, 10, 30, 100, 300, 1000, 3000, 10000). Used inside SingleData(), CombinationData().
}
\examples{
SimulateData()

}
\keyword{Simulation}
