% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationFct.R
\name{SimulateExperiment}
\alias{SimulateExperiment}
\title{SimulateExperiment() Function}
\usage{
SimulateExperiment(
  Nexperiment,
  paramA,
  paramB,
  paramC1,
  paramC2,
  paramC3,
  lambda1,
  lambda2,
  lambda3,
  rayA,
  rayB,
  ray1,
  ray2,
  ray3,
  Rep,
  sd
)
}
\arguments{
\item{Nexperiment}{: total number of experiments}

\item{paramA}{: Vectors containing parameters(Emin, Emax, m, IC50) of dose A}

\item{paramB}{: Vectors containing parameters(Emin, Emax, m, IC50) of dose B}

\item{paramC1}{: Vectors containing parameters(Emin, Emax, m, IC50) of combined dose}

\item{paramC2}{: Vectors containing parameters(Emin, Emax, m, IC50) of combined dose}

\item{paramC3}{: Vectors containing parameters(Emin, Emax, m, IC50) of combined dose}

\item{rayA}{: ray of dose A}

\item{rayB}{: ray of dose B}

\item{Rep}{: number of replicates}

\item{sd}{: sd (same variance for all rays)}

\item{rayC1}{: ray of combined dose}

\item{rayC2}{: ray of combined dose}

\item{rayC3}{: ray of combined dose}
}
\description{
This function does simulation with multiple experiments.(2 monotherapy data + 3 combination data)
}
\examples{
SimulateExperiment()
}
\keyword{Simulation}
