% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationFct.R
\name{SingleData}
\alias{SingleData}
\title{SingleData() Function}
\usage{
SingleData(param, Rep, sd, ray, doseName)
}
\arguments{
\item{param}{: Vectors containing parameters(Emin, Emax, m, IC50)}

\item{sd}{: sd}

\item{ray}{: following ray}

\item{doseName}{: name of the component}

\item{replicate}{: number of replicates}
}
\description{
This function does simulation with monotherapy data.
}
\examples{
SingleData()
}
\keyword{Simulation}
