% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TallaridaFct.R
\name{Tallaridamodel_Oneline}
\alias{Tallaridamodel_Oneline}
\title{Tallaridamodel_Oneline() Function}
\usage{
Tallaridamodel_Oneline(param, nameA, nameB, eff)
}
\arguments{
\item{param}{: Theoretical parameters of dose A and doseB}

\item{nameA}{: Name of dose A(lower potency dose)}

\item{nameB}{: Name of dose B(higher potency dose)}

\item{eff}{: Effects (often 50)}
}
\description{
This function calculates additivie isoboles by using Tallarida model (base-line dose is fixed to doseB, higher potency dose)
}
\examples{
Tallaridamodel_Oneline()
}
\keyword{Additive}
\keyword{Tallaridamodel,}
\keyword{isobole}
