% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TallaridaFct.R
\name{Tallaridamodel_Twoline}
\alias{Tallaridamodel_Twoline}
\title{Tallaridamodel_Twoline() Function}
\usage{
Tallaridamodel_Twoline(param, nameA, nameB, eff)
}
\arguments{
\item{param}{: Theoretical parameters of dose A and doseB}

\item{nameA}{: Name of dose A(lower potency dose)}

\item{nameB}{: Name of dose B(higher potency dose)}

\item{eff}{: Effects (often 50)}
}
\description{
This function calculates additivie isoboles by using Tallarida model that produce 2 lines of additivity. It should not be used as the 2 lines of additivity will have intersection points.
It is impossible for the pair of doses to produce 2 different additive effects. Only used to graphically represent the two additivities.
}
\examples{
Tallaridamodel_Twoline()

}
\keyword{Additive}
\keyword{Tallaridamodel,}
\keyword{isobole}
